7675: Update names to better reflect usage
