+++ java/deployment/cdk-artefact-custom-resources/src/main/java/sleeper/cdk/artefact/custom/CopyContainerImageLambda.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.cdk.artefact;
+package sleeper.cdk.artefact.custom;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.events.CloudFormationCustomResourceEvent;
