+++ java/deployment/cdk-artefact-custom-resources/src/test/java/sleeper/cdk/artefact/custom/testutil/NexusRepositoryStartupCheckStrategy.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.cdk.artefact.testutil;
+package sleeper.cdk.artefact.custom.testutil;
 
 import com.github.dockerjava.api.DockerClient;
 import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
